Adding consideration for float room numbers
I modified the existing code to consider that float room numbers are a possibility.

This meant changing the data type in a few places

+++ app/src/main/java/com/droidhats/campuscompass/viewmodels/SearchViewModel.kt
@@ -143,7 +143,7 @@ class SearchViewModel(application: Application) : AndroidViewModel(application) 
        }
     }
 
-    fun setIndoorDirections(startAndEnd: Pair<Int, Int>) {
+    fun setIndoorDirections(startAndEnd: Pair<Float, Float>) {
         IndoorNavigationRepository.getInstance().setStartAndEnd(startAndEnd)
     }
 }
